Initial work on a Kademlia implementation for Python 3 asyncio.
Implemented a simple RPC-over-UDP system as the basis for the Kademlia
protocol, the heart of which are the 4 RPCs `ping`, `store`, `find_node`,
and `find_value`.  A naive implemenation of `ping`, `store` as well as
the naive shortcut case of `find_value` are included in this.

Note that with this commit, we've got the basic plumbing for RPCs, but
none of the important hashtable implementation.

Also, setup.py and README.md are empty ;)