Building a rough verison of the core iterative `lookup_node` procedure, which appears to be somewhat working.  As of this version, I can store values in the network (manually with the direct RPC calls), use `lookup_node` to find the closest node to a key, then use RPC to retrieve that value.
Conceivably, the iterative `put` and `get` will be trivial to implement on
top of this.

Refactored some common operational things into the main package, and changed
`__main__` to always run a single standalone server, added `local_network.py`
as a self-contained 10-node network, and `local_client.py` which starts a node
and launches an embedded IPython for interactive testing.

Note, the `local_client.py` script does *not* (yet) join the local_network, as
this is part of what I'm still trying to build.